fix crash

+++ src/core/DynaRec_x64/emitter.h
@@ -65,7 +65,7 @@ struct Emitter final : public CodeGenerator {
     // Uses lea if the value is non-zero, or mov otherwise
     void moveAndAdd(Xbyak::Reg32 dest, Xbyak::Reg32 source, uint32_t value) {
         if (value != 0) {
-            lea(dest, dword[source.cvt64() + value]);
+            lea(dest, dword[source.cvt64() + static_cast<int32_t>(value)]);
         } else {
             mov(dest, source);
         }
